pyproject.toml
[project]
name = "python-mcp-demos"
version = "0.1.0"
description = "Demonstration of Python FastMCP servers"
readme = "README.md"
requires-python = "==3.13.*"
dependencies = [
    "fastmcp>=2.12.5",
    "debugpy>=1.8.0",
    "langchain-core>=0.3.0",
    "mcp>=1.3.0",
    "azure-identity>=1.25.1",
    "langchain==1.0.0a5",
    "langchain-openai>=1.0.1",
    "langchain-mcp-adapters>=0.1.11",
    "azure-ai-agents>=1.1.0",
    "agent-framework>=1.0.0b251016",
    "azure-cosmos>=4.9.0",
    "azure-monitor-opentelemetry>=1.6.4",
    "opentelemetry-instrumentation-starlette>=0.49b0",
    "logfire>=4.15.1",
    "azure-core-tracing-opentelemetry>=1.0.0b12"
]

[dependency-groups]
dev = [
    "ruff>=0.8.0",
    "pre-commit>=4.0.0",
]

[tool.ruff]
line-length = 120
target-version = "py310"
lint.select = ["E", "F", "I", "UP"]
lint.ignore = ["D203"]
